+++ packages/passkeys/passkeys/example/lib/pages/sign_in_page.dart
@@ -25,12 +25,22 @@ class _SignInPageState extends ConsumerState<SignInPage> {
     WidgetsBinding.instance.addPostFrameCallback((_) {
       final authService = ref.watch(authServiceProvider);
 
+      authService.getAvailability().then((value) {
+        debugPrint('passkey support: ${value.hasPasskeySupport}');
+        debugPrint(
+            'isUserVerifyingPlatformAuthenticatorAvailable: '
+                '${value.isUserVerifyingPlatformAuthenticatorAvailable}');
+        debugPrint(
+            'isConditionalMediationAvailable: '
+                '${value.isConditionalMediationAvailable}');
+        debugPrint('isNative: ${value.isNative}');
+      });
       // As soon as the view has been loaded prepare the autocompleted passkey sign in.
       authService
           .loginWithPasskeyConditionalUI()
           .then((value) => context.go(Routes.profile))
           .onError(
-        (error, stackTrace) {
+            (error, stackTrace) {
           if (error is PasskeyAuthCancelledException) {
             debugPrint(
                 'user cancelled authentication. This is not a problem. It can just be started again.');
@@ -86,7 +96,10 @@ class _SignInPageState extends ConsumerState<SignInPage> {
           if (error.value != null)
             Text(
               error.value!,
-              style: TextStyle(color: Theme.of(context).colorScheme.error),
+              style: TextStyle(color: Theme
+                  .of(context)
+                  .colorScheme
+                  .error),
             )
           else
             Container(),
@@ -122,7 +135,9 @@ class _SignInPageState extends ConsumerState<SignInPage> {
               style: OutlinedButton.styleFrom(
                 padding: const EdgeInsets.symmetric(vertical: 10),
                 side:
-                    BorderSide(width: 2, color: Theme.of(context).primaryColor),
+                BorderSide(width: 2, color: Theme
+                    .of(context)
+                    .primaryColor),
               ),
               onPressed: () => context.go(Routes.signUp),
               child: const Text('I want to create a new account'),