+++ packages/passkeys/passkeys_platform_interface/lib/types/types.dart
@@ -1,6 +1,7 @@
 export 'authenticate_request.dart';
 export 'authenticate_response.dart';
 export 'authenticator_selection.dart';
+export 'availability.dart';
 export 'mediation.dart';
 export 'register_request.dart';
 export 'register_response.dart';